requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
outcome==1.2.0
pandas==2.0.3
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
yarl==1.9.2
